pyproject.toml
[tool.poetry]
name = "rates"
version = "0.1.0"
description = "API to store and search for market financial quotes"
authors = ["Lucas João <lucasljoliveira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
markdown = "^3.5.1"
requests = "^2.31.0"
prettyconf = "^2.2.1"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
pytest = "^7.4.3"
pre-commit = "^3.5.0"
pytest-django = "^4.7.0"
factory-boy = "^3.3.0"
pytest-cov = "^4.1.0"
freezegun = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "rates.settings"
